renamed magntion-sdk file, now there's no magntion name in the sdk

+++ examples/sdk-deadlines/Main/MainReactor.cc
@@ -6,7 +6,7 @@ void MainReactor::construction() {
 
     slow = std::make_unique<NodeReactor>("slow", this);
 
-    for (size_t __lf_idx = 0; __lf_idx < parameters.n_fast.value; __lf_idx++) {
+    for (int __lf_idx = 0; __lf_idx < parameters.n_fast.value; __lf_idx++) {
         std::string __lf_inst_name = "fast_" + std::to_string(__lf_idx);
         fast.emplace_back(std::make_unique<NodeReactor>(__lf_inst_name, this));
     }
